+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+	"compilerOptions": {
+		"baseUrl": "src"
+	},
+	"include": [
+		"src"
+	]
+}
