+++ src/assets/img/dish7.png
