+++ src/assets/index.js
@@ -1,0 +1,25 @@
+import dish1 from './img/dish1.png';
+import dish2 from './img/dish2.png';
+import dish3 from './img/dish3.png';
+import dish4 from './img/dish4.png';
+import dish5 from './img/dish5.png';
+import dish6 from './img/dish6.png';
+import dish7 from './img/dish7.png';
+import dish8 from './img/dish8.png';
+import background from './img/cuate.png';
+import burger from './img/burger1.png';
+import avatar from './img/avatar.png';
+
+export {
+  dish1,
+  dish2,
+  dish3,
+  dish4,
+  dish5,
+  dish6,
+  dish7,
+  dish8,
+  background,
+  burger,
+  avatar,
+};
