+++ src/components/CartItem/style.scss
@@ -1,0 +1,76 @@
+.cart-item {
+	margin-top: 20px;
+	box-sizing: border-box;
+	border-top: 2px solid rgb(233, 233, 233);
+	display: flex;
+	padding-top: 30px;
+	width: 100%;
+	justify-content: space-between;
+	align-items: center;
+	&__details {
+		display: flex;
+		&-img {
+			width: 25%;
+			height: 25%;
+			margin-right: 5%;
+		}
+		&-info {
+			display: flex;
+			flex-direction: column;
+			justify-content: space-between;
+			width: 30%;
+		}
+		& .des, & .user {
+			font-size: 10px;
+			bottom: 10px;
+			font-family: 'Lexend Deca';
+			font-style: normal;
+			font-weight: 300;
+			font-size: 10px;
+			line-height: 20px;
+			color: #929292;
+		}
+		& .des {
+			box-sizing: border-box;
+		}
+	}
+	&__add {
+		position: relative;
+		top: 10px;
+		right: 5%;
+		display: flex;
+		flex-direction: row;
+		justify-content: space-between;
+		width: 30%;
+		&-adjust {
+			display: flex;
+			flex-direction: column;
+			&-qty {
+				position: relative;
+				bottom: 10px;
+				background-color: #F7F8FA;
+				display: flex;
+				justify-content: center;
+				align-items: center;
+				justify-content: space-between;
+				padding: 10px 10px;
+				border-radius: 8px;
+				width: 80%;
+			}
+			& i {
+				font-size: 15px;
+				cursor: pointer;
+				color: #1AC073;
+			}
+		}
+		& button {
+			border: none;
+			background-color: white;
+			color: red;
+			cursor: pointer;
+			&:hover {
+				opacity: 0.6;
+			}
+		}
+	}
+}
