+++ src/components/Navbar/Nav.js
@@ -1,19 +1,19 @@
-import "./Nav.scss";
-import { useNavigate } from "react-router-dom";
+import './style.scss';
+import { useNavigate } from 'react-router-dom';
 
 function AuthNav({ bot }) {
-	const navigate = useNavigate();
-	const handleClick = () => {
-		navigate("/");
-	};
+  const navigate = useNavigate();
+  const handleClick = () => {
+    navigate('/');
+  };
 
-	return (
-		<div style={{ bottom: bot ? 0 : null }} className='navbar'>
-			<p onClick={handleClick} className='navbar__logo'>
+  return (
+    <div style={{ bottom: bot ? 0 : null }} className='navbar'>
+      <p onClick={handleClick} className='navbar__logo'>
 				TBayEAT
-			</p>
-		</div>
-	);
+      </p>
+    </div>
+  );
 }
 
 export default AuthNav;
