--- src/components/Navbar/Nav.scss
@@ -1,19 +1,0 @@
-$primary-color: #1AC073;
-
-.navbar {
-	background-color: $primary-color;
-	position: relative;
-	width: 100%;
-	height: 80px;
-}
-
-.navbar__logo {
-	position: absolute;
-	left: 130px;
-	top: 20px;
-	color: white;
-	font-family: 'Squada One', cursive;
-	font-weight: 400;
-	font-size: 32px;
-	cursor: pointer;
-}
