+++ src/components/Navbar/style.scss
@@ -51,3 +51,20 @@ $sub-color: #F7F8FA;
 		}
 	}
 }
+
+.navbar {
+	background-color: $primary-color;
+	position: relative;
+	width: 100%;
+	height: 80px;
+	&__logo {
+		position: absolute;
+		left: 130px;
+		top: 20px;
+		color: white;
+		font-family: 'Squada One', cursive;
+		font-weight: 400;
+		font-size: 32px;
+		cursor: pointer;
+	}
+}
