--- src/components/SignIn/SignInPage.js
@@ -1,13 +1,0 @@
-import SignInForm from "./SignInForm";
-import AuthNav from "../Navbar/Nav";
-
-function SignIn() {
-	return (
-		<>
-			<AuthNav />
-			<SignInForm />
-		</>
-	);
-}
-
-export default SignIn;
