--- src/components/SignUp/SignUpPage.js
@@ -1,13 +1,0 @@
-import AuthNav from "../Navbar/Nav";
-import SignUpForm from "./SignUpForm";
-
-function SignUpPage() {
-	return (
-		<>
-			<AuthNav />
-			<SignUpForm />
-		</>
-	);
-}
-
-export default SignUpPage;
