+++ src/pages/Login/SignInPage.js
@@ -1,0 +1,13 @@
+import SignInForm from './SignInForm';
+import AuthNav from 'components/Navbar/Nav';
+
+function SignIn() {
+  return (
+    <>
+      <AuthNav />
+      <SignInForm />
+    </>
+  );
+}
+
+export default SignIn;
