+++ src/pages/Payment/style.scss
@@ -1,0 +1,226 @@
+$primary-color: #1AC073;
+$sub-color: #F7F8FA;
+
+.home__top-menu {
+	background-color: $primary-color;
+	height: 66.5vh;
+}
+
+.payment {
+	height: 100%;
+	padding: 100px 120px;
+	&__current-page {
+		background-color: $sub-color;
+		width: 100%;
+		height: 100px;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+		font-family: 'ABeeZee';
+		font-style: normal;
+		font-weight: 400;
+		font-size: 18px;
+		line-height: 21px;
+		color: #1AC073;
+		cursor: pointer;
+	}
+	&__details {
+		padding: 0 30px;
+		margin-top: 40px;
+		position: relative;
+		display: flex;
+		justify-content: space-between;
+		font-family: 'Lexend Deca';
+		font-style: normal;
+		font-weight: 500;
+		font-size: 16px;
+		line-height: 20px;
+		color: #22262A;
+		&-qty {
+			position: relative;
+			left: 12vw;
+		}
+		&-price {
+			position: relative;
+			left: 25.5vw;
+		}
+		&-unit {
+			position: relative;
+			right: 4vh;
+		}
+	}
+	&__line {
+		margin-top: 50px;
+		width: 100%;
+		height: 2px;
+		background-color: rgb(233, 233, 233);
+	}
+	&__total {
+		margin-top: 50px;
+		display: flex;
+		justify-content: space-between;
+		&-voucher {
+			display: flex;
+			justify-content: center;
+			align-items: center;
+			& input[type="text"] {
+				font-size: 20px;
+				padding: 19px;
+				border: 1px solid rgb(233, 233, 233);
+				&:focus {
+					outline: none;
+				}
+			}
+			& button {
+				font-family: 'Lexend Deca';
+				font-style: normal;
+				font-weight: 500;
+				font-size: 18px;
+				align-items: center;
+				color: #FFFFFF;
+				border: none;
+				background-color: #1AC073;
+				padding: 20px;
+				border-radius: 0 7px 7px 0;
+				cursor: pointer;
+				&:hover {
+					opacity: 0.8;
+				}
+			}
+		}
+		&-result {
+			display: flex;
+			flex-direction: column;
+			width: 30%;
+			font-family: 'Poppins';
+			font-style: normal;
+			font-weight: 400;
+			font-size: 18px;
+			color: #262626;
+			&-sub, &-ship, &-coupon, &-total {
+				display: flex;
+				justify-content: space-between;
+				margin-top: 30px;
+			}
+			&-total {
+				margin-top: 60px;
+				padding-top: 30px;
+				margin-bottom: 30px;
+				border-top: 2px solid rgb(233, 233, 233);
+				font-family: 'Lexend Deca';
+				font-style: normal;
+				font-weight: 500;
+				font-size: 30px;
+				color: #22262A;
+			}
+			&-btn1, &-btn2 {
+				margin-top: 30px;
+				font-family: 'Lexend Deca';
+				font-style: normal;
+				font-weight: 400;
+				font-size: 18px;
+				line-height: 20px;
+				color: #1AC073;
+				padding: 20px;
+				border-radius: 10px;
+				border: none;
+				box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.3);
+				cursor: pointer;
+				&:hover {
+					opacity: 0.8;
+				}
+			}
+			&-btn1 {
+				background-color: white;
+				border: 1px solid #1AC073;
+			}
+			&-btn2 {
+				background-color: #1AC073;
+				color: white;
+			}
+		}
+	}
+	&__charge {
+		position: absolute;
+		background: rgba(34, 34, 34, 0.3);
+		top: 0;
+		width: 100%;
+		height: 268%;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+		&-form {
+			width: 60%;
+			height: 30%;
+			padding: 40px;
+			background-color: white;
+			border-radius: 30px;
+			&-icon {
+				display: flex;
+				justify-content: space-between;
+				font-size: 30px;
+				color: #1AC073;
+				cursor: pointer;
+				&:hover {
+					opacity: 0.8;
+				}
+			}
+			&-title {
+				margin-top: 10px;
+				text-align: center;
+				font-family: 'ABeeZee';
+				font-style: normal;
+				font-weight: 400;
+				font-size: 32px;
+				line-height: 38px;
+				color: #1AC073;
+			}
+			&-current {
+				display: flex;
+				justify-content: center;
+				margin-top: 30px;
+				&-page {
+					display: flex;
+					justify-content: space-between;
+					width: 180px;
+					font-family: 'Lexend Deca';
+					font-style: normal;
+					font-weight: 700;
+					font-size: 18px;
+					line-height: 22px;
+					letter-spacing: -0.02em;
+					color: #FFFFFF;
+					text-shadow: 0px 0px 2px rgba(0, 0, 0, 0.25);
+					& .page {
+						background-color: #DFDEDE;
+						padding: 9px 15px;
+						border-radius: 50%;
+					}
+					& .line {
+						position: relative;
+						top: 20px;
+						background-color: #DFDEDE;
+						height: 2px;
+						width: 100%;
+					}
+				}
+			}
+			&-info {
+				margin-top: 30px;
+				display: flex;
+				justify-content: space-between;
+				background-color: red;
+				padding: 0 30px;
+				&-left {
+					display: flex;
+					flex-direction: column;
+					height: 300px;
+				}
+				&-right {
+					display: flex;
+					flex-direction: column;
+				}
+			}
+		}
+	}
+}
