+++ src/pages/Profile/style.scss
