+++ src/pages/Signin/SignUpPage.js
@@ -1,0 +1,13 @@
+import AuthNav from 'components/Navbar/Nav';
+import SignUpForm from './SignUpForm';
+
+function SignUpPage() {
+  return (
+    <>
+      <AuthNav />
+      <SignUpForm />
+    </>
+  );
+}
+
+export default SignUpPage;
